no need for data folder all contained in gatsby-config.js now
