--- data/siteConfig.js
@@ -1,15 +1,0 @@
-module.exports = {
-  siteTitle: 'scottspence.me',
-  siteUrl: 'https://scottspence.me',
-  nameContent: 'Scott Spence - portfolio',
-  developerName: 'Scott Spence',
-  developerUrl: 'https.scottspence.me',
-  descriptionContent: 'Scott Spence - web developer',
-  keywordsContent:
-    'web developer, javascript, react, portfolio, information',
-  imageLink: 'https://scottspence.me/icons/icon-512x512.png',
-  faviconPng: './src/images/favicon.png',
-  backgroundColour: '#663399', // this is for favicon and manifest
-  themeColour: '#755f9f', // this is for favicon and manifest
-  siteLanguage: 'en-GB'
-}
